reorganize errors
