+++ backend/src/domain/repositories/recipe/errors/get.rs
@@ -1,0 +1,46 @@
+use eyre::eyre;
+use sqlx::Error as SQLXError;
+use std::sync::PoisonError;
+use thiserror::Error;
+use uuid::Uuid;
+
+use crate::domain::entities::recipe::errors::ValidationError;
+
+#[derive(Error, Debug)]
+pub enum GetRecipeByIdError {
+    #[error("The recipe with ID of {0} was not found")]
+    NotFound(Uuid),
+
+    #[error(transparent)]
+    ValidationError(#[from] ValidationError),
+
+    #[error(transparent)]
+    UnknownError(#[from] eyre::Error),
+}
+
+impl GetRecipeByIdError {
+    pub fn with_id(id: &Uuid, e: SQLXError) -> Self {
+        match e {
+            SQLXError::RowNotFound => Self::NotFound(*id),
+            _ => Self::UnknownError(e.into()),
+        }
+    }
+}
+
+impl<T> From<PoisonError<T>> for GetRecipeByIdError {
+    fn from(_value: PoisonError<T>) -> Self {
+        eyre!("Recipe repository lock was poisoned during a previous access and can no longer be locked").into()
+    }
+}
+
+impl From<SQLXError> for GetRecipeByIdError {
+    fn from(e: SQLXError) -> Self {
+        Self::UnknownError(e.into())
+    }
+}
+
+impl From<serde_json::Error> for GetRecipeByIdError {
+    fn from(e: serde_json::Error) -> Self {
+        Self::UnknownError(e.into())
+    }
+}
